requirements.txt
litellm>=1.10.1
python-chess>=1.999
pyyaml>=6.0.1
python-dotenv>=1.0.0 
chess~=1.11.1
pandas~=2.2.3
matplotlib~=3.9.3
seaborn~=0.13.2